requirements.txt
Pillow
beeprint
numpy
